% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.explain.code.R
\name{text.explain.code}
\alias{text.explain.code}
\title{Generate code explanation using the Generative AI model based on a prompt}
\usage{
text.explain.code(model.parameter, temperature, prompt, language = "R")
}
\arguments{
\item{model.parameter}{A character vector containing the Generative AI service provider,
corresponding model, version, API key, and proxy status.}

\item{temperature}{A numeric value. A higher value results in more creative responses,
while a lower value produces more straightforward text.}

\item{prompt}{A character string representing the query for text generation.}

\item{language}{A character string representing the programming language of the code in
the prompt.}
}
\value{
If successful, a character string (code explanation) generated by the Generative AI
model based on the provided query and parameters. If the API response indicates an error,
the function halts execution and provides an error message.
}
\description{
This function establishes a connection to a Generative AI model by providing essential
parameters and generates a code explanation based on the prompt.
}
\details{
Providing accurate and valid information for each parameter is crucial
to ensure successful text generation by the Generative AI model. If any of the
provided parameters is incorrect, the function will respond with an error message based
on the information received from the API. Use the function \code{available.models} to
see all supported Generative AI models.

A complete prompt behind the function is as follows:

Explain the following \code{language} code:

# Code starts #

\code{prompt}

# Code ends #
}
\examples{
\dontrun{
 # Get available models
 models = available.models()

 # Connect to the model, replace API_KEY with your api key
 google.model = connect.genai("google",
                              models$google$model[1],
                              models$google$version[1],
                              "API_KEY",
                              FALSE)

 # Connect to the model, replace API_KEY with your api key
 openai.model = connect.genai("openai",
                              models$openai$model[1],
                              models$openai$version[1],
                              "API_KEY",
                              FALSE)

 # Generate text
 temperature = 0.9
 prompt = "foo <- function(n) {
             if (n <= 0) {
               return(0)
             } else if (n == 1) {
               return(1)
             } else {
               return(foo(n - 1) + foo(n - 2))
             }
           }"
 code.explanation = text.explain.code(google.model,
                                      temperature,
                                      prompt,
                                      language = "R")
 cat(code.explanation)

 code.explanation = text.explain.code(openai.model,
                                      temperature,
                                      prompt,
                                      language = "R")
 cat(code.explanation)
}

}
