% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.models.R
\name{available.models}
\alias{available.models}
\title{Get supported Generative AI models}
\usage{
available.models()
}
\value{
If successful, the function returns a list containing Generative AI
service providers and their corresponding models. If the function encounters an error,
it will halt execution and provide an error message.
}
\description{
This function sends a request to GenAI database API
to retrieve information about available Generative AI models.
}
\details{
The function utilizes the GenAI database API to fetch the latest information about
available Generative AI models. The retrieved data includes details about different models
offered by various service providers.
}
\examples{
\dontrun{
 models = available.models()
 models
}

}
